refactor: Member와 UserDetails 분리
